+++ src/game/nlcore/elements/script.ts
@@ -4,15 +4,19 @@ import {LogicAction} from "@core/action/logicAction";
 import {Actionable} from "@core/action/actionable";
 import {GameState} from "@player/gameState";
 import {Chained, Proxied} from "@core/action/chain";
-import type {Storable} from "@core/elements/persistent/storable";
+import type {Namespace, Storable} from "@core/elements/persistent/storable";
 import {ScriptAction} from "@core/action/actions/scriptAction";
 import {LiveGame} from "@core/game/liveGame";
+import { NameSpaceContent } from "./persistent/type";
+
+export type NamespaceGetter = <T extends NameSpaceContent<keyof T>>(namespace: string) => Namespace<T>;
 
 export interface ScriptCtx {
     gameState: GameState;
     game: Game;
     liveGame: LiveGame;
     storable: Storable;
+    $: NamespaceGetter;
 }
 
 type ScriptRun = (ctx: ScriptCtx) => ScriptCleaner | void;
@@ -22,12 +26,19 @@ export type ScriptCleaner = () => void;
 export class Script extends Actionable<object> {
     /**@internal */
     static getCtx({gameState}: { gameState: GameState }): ScriptCtx {
+        const liveGame = gameState.game.getLiveGame();
+        const storable = liveGame.getStorable();
         return {
             gameState,
             game: gameState.game,
-            liveGame: gameState.game.getLiveGame(),
-            storable: gameState.game.getLiveGame().getStorable(),
+            liveGame,
+            storable,
+            $: (namespace: string) => storable.getNamespace(namespace),
         };
+    }
+
+    public static execute(handler: ScriptRun): Proxied<Script, Chained<LogicAction.Actions>> {
+        return new Script(handler) as Proxied<Script, Chained<LogicAction.Actions>>;
     }
 
     /**@internal */
@@ -36,7 +47,14 @@ export class Script extends Actionable<object> {
     constructor(handler: ScriptRun) {
         super();
         this.handler = handler;
-        return this.chain() satisfies Proxied<Script, Chained<LogicAction.Actions>>;
+        
+        const chain = this.chain();
+        const action = new ScriptAction(
+            chain,
+            ScriptAction.ActionTypes.action,
+            new ContentNode<Script>().setContent(this)
+        );
+        return this.chain(action) satisfies Proxied<Script, Chained<LogicAction.Actions>>;
     }
 
     /**@internal */